Implement cursor based pagination
It is the most optimal way to paginate the huge dataset. We can move
forward or backward but we can't jump to a specific page with this
technique.

When requesting first page, each records will be returned
with a `_cursor` which is a `base64` string of ordering fields,
ordering directions and their values.

Next page can be fetched by taking `_cursor` from the last record.
Prev page cab be fetched by taking `_cursor` from the first record.